+++ lib/dart_di.dart
@@ -1,4 +1,4 @@
 library dart_di;
 
-export 'core/dart_di.dart';
+export 'core/dart_di.dart' show DDI;
 export 'features/ddi_interceptor.dart';