+++ test/clazz_test/father.dart
@@ -1,0 +1,7 @@
+import 'mother.dart';
+
+class Father {
+  Father({required this.mother});
+
+  final Mother mother;
+}