+++ test/clazz_test/mother.dart
@@ -1,0 +1,7 @@
+import 'father.dart';
+
+class Mother {
+  Mother({required this.father});
+
+  final Father father;
+}