+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Scopes: Singleton, Application, Dependent, Session and Widget.
+* Support register, get, dispose and destroy.
+* Qualifiers support.
+* Decorators support.
+* Interceptors support.
+* Conditional register.
+* Post Construct support.
+* Circular Dependency Injection Detector (Experimental).